Add files via upload
A loan application passes through a number of stages and it is still not sure whether it can be approved or declined. To reduce this time-consuming process, and to reduce the risk associated with it, various loan prediction models are introduced. 
The aim of this paper is to compare various Loan prediction models created using Machine learning techniques and to select the best one with the least error which can be used in real-world scenarios by banks or financial institutions to predict whether the loan application should be approved or not which further provides a quick and easy way to choose a deserving candidate from a number of applicants who have submitted their loan approval requests.